+++ src/aws-api-mcp-server/awslabs/aws_api_mcp_server/core/common/config.py
@@ -18,15 +18,15 @@ import tempfile
 from pathlib import Path
 
 
-def get_region() -> str:
-    """Get the default region for executing cli commands."""
-    if aws_region := os.getenv('AWS_REGION'):
-        return aws_region
+def get_region(profile_name: str | None = None) -> str:
+    """Get the region depending on configuration."""
+    if AWS_REGION:
+        return AWS_REGION
 
     fallback_region = 'us-east-1'
 
-    if AWS_API_MCP_PROFILE_NAME:
-        return boto3.Session(profile_name=AWS_API_MCP_PROFILE_NAME).region_name or fallback_region
+    if profile_name:
+        return boto3.Session(profile_name=profile_name).region_name or fallback_region
 
     return boto3.Session().region_name or fallback_region
 
@@ -56,8 +56,10 @@ def get_env_bool(env_key: str, default: bool) -> bool:
 
 FASTMCP_LOG_LEVEL = os.getenv('FASTMCP_LOG_LEVEL', 'WARNING')
 AWS_API_MCP_PROFILE_NAME = os.getenv('AWS_API_MCP_PROFILE_NAME')
-DEFAULT_REGION = get_region()
+AWS_REGION = os.getenv('AWS_REGION')
+DEFAULT_REGION = get_region(AWS_API_MCP_PROFILE_NAME)
 READ_OPERATIONS_ONLY_MODE = get_env_bool(READ_ONLY_KEY, False)
 OPT_IN_TELEMETRY = get_env_bool(TELEMETRY_KEY, True)
 WORKING_DIRECTORY = os.getenv('AWS_API_MCP_WORKING_DIR', get_server_directory() / 'workdir')
 REQUIRE_MUTATION_CONSENT = get_env_bool(REQUIRE_MUTATION_CONSENT_KEY, False)
+EMBEDDING_MODEL_DIR = os.getenv('EMBEDDING_MODEL_DIR', get_server_directory() / 'embedding_models')