+++ src/aws-api-mcp-server/awslabs/aws_api_mcp_server/core/common/helpers.py
@@ -15,11 +15,17 @@
 import json
 import os
 import re
+import requests
+import shutil
+import tempfile
 import time
+import zipfile
+from .config import EMBEDDING_MODEL_DIR
 from botocore.response import StreamingBody
 from contextlib import contextmanager
 from datetime import datetime
 from loguru import logger
+from pathlib import Path
 from typing import Any
 
 
@@ -70,3 +76,35 @@ def validate_aws_region(region: str):
         error_message = f'{region} is not a valid AWS Region'
         logger.error(error_message)
         raise ValueError(error_message)
+
+
+def download_embedding_model(model_name: str):
+    """Download embedding model from AWS."""
+    download_url = f'https://models.knowledge-mcp.global.api.aws/{model_name}.zip'
+
+    with tempfile.TemporaryDirectory() as tmp_dir:
+        logger.debug('Dowloading embedding model {} to {}', model_name, tmp_dir)
+        response = requests.get(download_url, stream=True, timeout=30)
+        response.raise_for_status()
+
+        zip_path = os.path.join(tmp_dir, f'{model_name}.zip')
+        Path(
+            zip_path
+        ).parent.mkdir()  # HF models are structured as '<org_name>/<model>' so we have to create the parent folder
+
+        with open(zip_path, 'wb') as f:
+            for chunk in response.iter_content(chunk_size=8192):
+                f.write(chunk)
+
+        extract_dir = os.path.join(EMBEDDING_MODEL_DIR, model_name)
+        logger.debug(
+            'Extracting embedding model from {} to {}', zip_path, os.path.abspath(extract_dir)
+        )
+        try:
+            with zipfile.ZipFile(zip_path, 'r') as zip_ref:
+                zip_ref.extractall(extract_dir)
+        except Exception as e:
+            logger.error('Failed to extract embedding model: {}', str(e))
+            if os.path.exists(extract_dir):
+                shutil.rmtree(extract_dir)
+            raise